+++ mods/tabs/mod.js
@@ -1,0 +1,19 @@
+/*
+ * tabs
+ * (c) 2020 dragonwocky <thedragonring.bod@gmail.com> (https://dragonwocky.me/)
+ * under the MIT license
+ */
+
+'use strict';
+
+// this is just a pseudo mod to "separate" the options
+// from the core module - the core still handles actually
+// making it work.
+module.exports = {
+  id: 'e1692c29-475e-437b-b7ff-3eee872e1a42',
+  tags: ['core', 'extension'],
+  name: 'tabs',
+  desc: 'have multiple notion pages open in a single window.',
+  version: '0.1.0',
+  author: 'dragonwocky',
+};